+++ lib/rdf/rdfa.rb
@@ -9,7 +9,7 @@ module RDF
   #   require 'rdf/rdfa'
   #
   # @example Parsing RDF statements from an XHTML+RDFa file
-  #   RDF::RDFa::Reader.open("etc/foaf.html") do |reader|
+  #   RDF::RDFa::Reader.open("etc/doap.html") do |reader|
   #     reader.each_statement do |statement|
   #       puts statement.inspect
   #     end
@@ -27,7 +27,7 @@ module RDF
     require 'rdf/rdfa/vocab'
     require 'rdf/rdfa/patches/string_hacks'
     autoload :Expansion,  'rdf/rdfa/expansion'
-    autoload :Profile,    'rdf/rdfa/profile'
+    autoload :Context,    'rdf/rdfa/context'
     autoload :Reader,     'rdf/rdfa/reader'
     autoload :Writer,     'rdf/rdfa/writer'
     autoload :VERSION,    'rdf/rdfa/version'